Test mm to inches
