+++ WeatherXMTests/Toolkit/UnitsConverterTests.swift
@@ -29,6 +29,22 @@ final class UnitsConverterTests: XCTestCase {
 		}
     }
 
+	func testMMtoInched() throws {
+		let expectedResults: [Double: Double] = [-10.0: -0.3937007874,
+												  -1.0: -0.0393700787,
+												  0.0: 0.0,
+												  1.0: 0.0393700787,
+												  5.0: 0.1968503937,
+												  10.8: 0.4251968504,
+												  100.0: 3.937007874,
+												  1000.0: 39.3700787402]
+
+		expectedResults.forEach { millimeters, inches in
+			let result =  unitsConverter.millimetersToInches(mm: millimeters).rounded(toPlaces: 10)
+			XCTAssert(result == inches, "The result is \(result), it should be \(inches) for \(millimeters)")
+		}
+	}
+
     func testPerformanceExample() throws {
         // This is an example of a performance test case.
         self.measure {
